Phase 2: move mitigation.py/healing.py into dense_evolution.mitigation/ (zne.py + healing.py)
- git mv dense_evolution/mitigation.py -> dense_evolution/mitigation/zne.py
  (renamed, ZNE-specific vocabulary vs. the shared "mitigation" subpackage name)
- git mv dense_evolution/healing.py -> dense_evolution/mitigation/healing.py
- new mitigation/__init__.py re-exports both modules' public API
- zne.py's internal `from .healing import calculate_delta_preemp` needed no
  change: both modules move into mitigation/ together in this same PR, so
  the relative import still resolves correctly (same situation as gates.py's
  .registry import in the circuits/ PR)

Backward compatibility, two different cases in the same PR:
- dense_evolution.healing gets a real shim: its dotted path genuinely
  changes to dense_evolution.mitigation.healing (same category as physics/,
  backends/).
- dense_evolution.mitigation needs NO separate shim file: the new
  subpackage is named exactly "mitigation", matching the old top-level
  module's own name, so dense_evolution.mitigation as a dotted path is
  unchanged by construction (same category as interop/).

Self-caught mistake, found and fixed during verification: initially created
a dense_evolution/mitigation.py shim file anyway (reasoning it through
wrong -- assumed "renamed file" implies "needs a shim" without checking
the specific case). The new shim-coverage-check discipline from PR #83
caught it immediately: that shim file showed 0% coverage even though
tests/test_mitigation.py imports through exactly that dotted path.
Investigation confirmed Python always resolves an import to a same-named
PACKAGE over a plain module file in the same location -- so the shim file
was completely dead, unreachable code from the moment it was created.
Deleted it; dense_evolution/mitigation/__init__.py already covers that
exact dotted path.

Fixed one real regression, same private-helper-import pattern already
seen in interop/, circuits/, backends/: tests/test_mitigation.py imported
three private JIT-core helpers (_zero_noise_extrapolation_healing_core,
_zne_density_matrix_core, _uhlmann_fidelity_core) from the
dense_evolution.mitigation path at three separate call sites -- pointed
all three at dense_evolution.mitigation.zne directly instead.

Verified dense_evolution.healing's shim IS genuinely exercised (100%
coverage): tests/test_healing.py does `from dense_evolution import
healing` (bare submodule import) and calls public functions via
`healing.<name>(...)`, and tools/ia_utils/vector_healing.py +
adversarial_vector_attack.py both import from dense_evolution.healing
directly -- no coverage-gap fix needed here, unlike backends/'s mps.py.

Full suite: 804 passed, 17 skipped, 0 failed.

Part of #76.
