+++ dense_evolution/__init__.py
@@ -14,7 +14,7 @@ from .interop import (
 )
 from .autodiff import circuit_to_energy_fn
 from .backends.mps import MPSSimulator
-from .mitigation import (richardson_extrapolate, zero_noise_extrapolation, polynomial_extrapolate,
+from .mitigation.zne import (richardson_extrapolate, zero_noise_extrapolation, polynomial_extrapolate,
                           project_to_physical, uhlmann_fidelity, zne_density_matrix,
                           jsd_predictive_zne_density_matrix,
                           richardson_extrapolate_jit, zero_noise_extrapolation_jit,
