+++ dense_evolution/mitigation/__init__.py
@@ -1,0 +1,25 @@
+"""Mitigation subpackage: Zero-Noise Extrapolation (zne) and the
+predictive-healing primitives (healing) it composes."""
+from .zne import (
+    richardson_extrapolate, zero_noise_extrapolation, polynomial_extrapolate,
+    project_to_physical, uhlmann_fidelity, zne_density_matrix,
+    jsd_predictive_zne_density_matrix,
+    richardson_extrapolate_jit, zero_noise_extrapolation_jit,
+    polynomial_extrapolate_jit, uhlmann_fidelity_jit, zne_density_matrix_jit,
+)
+from .healing import (
+    calculate_advanced_sigma, calculate_phi_ab, calculate_vettore_dinamico,
+    calculate_vettore_statico, calculate_delta_preemp, evaluate_phi_trigger,
+    calculate_jax_reflection, MemoryReflectionEngine, GLOBAL_CONSTANTS,
+)
+
+__all__ = [
+    "richardson_extrapolate", "zero_noise_extrapolation", "polynomial_extrapolate",
+    "project_to_physical", "uhlmann_fidelity", "zne_density_matrix",
+    "jsd_predictive_zne_density_matrix",
+    "richardson_extrapolate_jit", "zero_noise_extrapolation_jit",
+    "polynomial_extrapolate_jit", "uhlmann_fidelity_jit", "zne_density_matrix_jit",
+    "calculate_advanced_sigma", "calculate_phi_ab", "calculate_vettore_dinamico",
+    "calculate_vettore_statico", "calculate_delta_preemp", "evaluate_phi_trigger",
+    "calculate_jax_reflection", "MemoryReflectionEngine", "GLOBAL_CONSTANTS",
+]
